+++ core/tests/tcp_socket_test.cpp
@@ -72,7 +72,7 @@ protected:
             for (;;)
             {
                 // echo received data.
-                int nread = sock.read (buf, sizeof (buf));
+                ssize_t nread = sock.read (buf, sizeof (buf));
                 if (nread == -1)
                 {
                     if (join::lastError == Errc::TemporaryError)
