+++ core/tests/udp_socket_test.cpp
@@ -70,7 +70,7 @@ protected:
         if (buffer)
         {
             Udp::Endpoint from;
-            int nread = _server.readFrom (buffer.get (), _server.canRead (), &from);
+            ssize_t nread = _server.readFrom (buffer.get (), _server.canRead (), &from);
             if (nread >= 0)
             {
                 _server.writeTo (buffer.get (), nread, from);
